% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_uncal_images.R
\name{find_uncal_photos}
\alias{find_uncal_photos}
\title{Find uncalibarated images}
\usage{
find_uncal_photos(proj_name, proj_dir, img_dir = proj_dir)
}
\arguments{
\item{proj_name}{Name of project}

\item{proj_dir}{Path to the project}

\item{img_dir}{Path to the folder holding the raw images}
}
\value{
A vector of image names
}
\description{
Find uncalibarated images
}
\examples{
\dontrun{
test <- find_uncal_photos(proj_name = "lake_ellsmere_2019",
                          proj_dir  = '../New_Zealand/Processing/Lake_Ellsmere',
                          img_dir   = '../New_Zealand/Raw_Photos/Lake_Ellsmere')
}
}
