% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_uncal_images.R
\name{handle_uncal_photos}
\alias{handle_uncal_photos}
\title{Move uncalibrated images to a separate folder or delete them}
\usage{
handle_uncal_photos(img_paths, move_to, delete = FALSE)
}
\arguments{
\item{img_paths}{the paths to the images to move or delete}

\item{move_to}{if delete is false, where to move the images}

\item{delete}{TRUE or FALSE}
}
\value{
New image paths, or TRUE if deleted
}
\description{
Move uncalibrated images to a separate folder or delete them
}
